Add support for custom typed values
This commit add an option to allow values outside the options to be
commited. In my use case, I just want to show options as suggestions so
the user should be able to use whatever he typed as the value if he
wants to.

Also, send the `this.label` so that we it makes it easier to pattern
match the event in LiveView insteaf of using the `input[name]`. For
instance:

```Elixir
def handle_event("autocomplete-commit", %{"label" => "my-label", "value" => value}, socket) do
  ...
end
```